Remove withStyles from PopinManager
