+++ components/popin-manager/index.ts
@@ -1,0 +1,1 @@
+export { default as PopinManager } from "./PopinManager";
